### 2022-04-04 #169 @whedon
[PRE REVIEW]: SNIK Graph—Visualizing Knowledge about Management of Hospital Information Systems
**Submitting author:** @KonradHoeffner (<a href="http://orcid.org/0000-0001-7358-3217">Konrad Höffner</a>)
**Repository:** <a href="https://github.com/snikproject/snik-graph" target ="_blank">https://github.com/snikproject/snik-graph</a>
**Version:** 3.0.0
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
