### 2022-04-07 68f6913
Explain different deploys in README.md.
