+++ assets/main.34f00dab.js
@@ -1,182 +1,0 @@
-var et=Object.defineProperty,tt=Object.defineProperties;var nt=Object.getOwnPropertyDescriptors;var Ce=Object.getOwnPropertySymbols;var st=Object.prototype.hasOwnProperty,ot=Object.prototype.propertyIsEnumerable;var ce=(s,e,t)=>e in s?et(s,e,{enumerable:!0,configurable:!0,writable:!0,value:t}):s[e]=t,O=(s,e)=>{for(var t in e||(e={}))st.call(e,t)&&ce(s,t,e[t]);if(Ce)for(var t of Ce(e))ot.call(e,t)&&ce(s,t,e[t]);return s},le=(s,e)=>tt(s,nt(e));var b=(s,e,t)=>(ce(s,typeof e!="symbol"?e+"":e,t),t);/* empty css              */import{l as c,S as at,c as re,a as N,t as ke,b as it,d as rt,G as ct,s as lt,h as dt,F as ut,N as ht,e as de}from"./vendor.9308ee05.js";const gt=function(){const e=document.createElement("link").relList;if(e&&e.supports&&e.supports("modulepreload"))return;for(const o of document.querySelectorAll('link[rel="modulepreload"]'))n(o);new MutationObserver(o=>{for(const i of o)if(i.type==="childList")for(const a of i.addedNodes)a.tagName==="LINK"&&a.rel==="modulepreload"&&n(a)}).observe(document,{childList:!0,subtree:!0});function t(o){const i={};return o.integrity&&(i.integrity=o.integrity),o.referrerpolicy&&(i.referrerPolicy=o.referrerpolicy),o.crossorigin==="use-credentials"?i.credentials="include":o.crossorigin==="anonymous"?i.credentials="omit":i.credentials="same-origin",i}function n(o){if(o.ep)return;o.ep=!0;const i=t(o);fetch(o.href,i)}};gt();const m={defaultSubOntologies:["bb","ob","ciox","he","it4it"],allSubOntologies:["meta","bb","ob","ciox","he","he-unconsolidated","it4it"],color:new Map([["ciox","rgb(80,255,250)"],["meta","rgb(255,80,80)"],["ob","rgb(255,173,30)"],["bb","rgb(30,152,255)"],["he","rgb(150,255,120)"],["it","rgb(204, 0, 204)"],["it4it","rgb(255, 255, 0)"]]),nodeSize:39,helperGraphs:["limes-exact","match"],activeOptions:[],searchCloseMatch:!0,logLevelConsole:"debug",logLevelDisplay:"info",logLevelMemory:"debug",layoutCacheMinRecall:.95,layoutCacheMinPrecision:.5,language:"en",download:{image:{max:{width:5e3,height:4e3},standard:{width:1920,height:1920}}},sparql:{endpoint:"https://www.snik.eu/sparql",graph:"http://www.snik.eu/ontology",instances:!1,isSnik:!0},multiview:{initialTabs:1,warnOnSessionLoad:!0}},Pe={GRAPH_BB:"http://www.snik.eu/ontology/bb",PREFIX:"http://www.snik.eu/ontology/",ENDPOINT:"https://www.snik.eu/sparql"};async function x(s,e,t=m.sparql.endpoint){const n=typeof window!="undefined";let o=t+"?query="+encodeURIComponent(s)+"&format=json";e&&(o+="&default-graph-uri="+encodeURIComponent(e));try{const r=(await(await fetch(o)).json()).results.bindings;return n&&console.groupCollapsed("SPARQL "+s.split(`
[…]
