+++ globals.d.ts
@@ -1,6 +1,3 @@
-import jQuery from "jquery";
-
-declare global {
-	//const log: log.RootLogger & { logs: Array<string> };
-	const JQuery: typeof jQuery;
-}
+//declare global {
+//const log: log.RootLogger & { logs: Array<string> };
+//}
