### 2022-04-14 747a75a
Resolve #16: Add taphold context menu event for touch screens.
