### 2022-04-19 92cde1e
Resolve #340: Display version from package.json.
