### 2022-04-19 cde1620
Resolve #384: Fix context menu tooltips.
