### 2022-05-19 a22bcc3
Adapt to ontology refactoring.
