+++ test/sparql.test.ts
@@ -14,24 +14,24 @@ const GRAPH_SNIK_META = "http://www.snik.eu/ontology/meta";
 describe("sparql", () => {
 	describe("#sparql()", () => {
 		it(`${GRAPH_GROUP_SNIK} should contain between ${EXPECTED_CLASSES_MIN} and ${EXPECTED_CLASSES_MAX} classes`, () => {
-			return sparql.select("select count(?class) as ?count {?class a owl:Class}", GRAPH_GROUP_SNIK).then((bindings) => {
+			return sparql.select("select count(?class) as ?count {?class a [rdfs:subClassOf meta:Top].}", GRAPH_GROUP_SNIK).then((bindings) => {
 				bindings[0].should.have.property("count");
 				parseInt(bindings[0].count.value).should.be.within(EXPECTED_CLASSES_MIN, EXPECTED_CLASSES_MAX);
 			});
 		});
 		it(`${GRAPH_SNIK_META} should contain between ${EXPECTED_META_CLASSES_MIN} and ${EXPECTED_META_CLASSES_MAX} classes`, () => {
-			return sparql.select("select count(?class) as ?count {?class a owl:Class}", GRAPH_SNIK_META).then((bindings) => {
+			return sparql.select("select count(?class) as ?count {?class a owl:Class.}", GRAPH_SNIK_META).then((bindings) => {
 				bindings[0].should.have.property("count");
 				parseInt(bindings[0].count.value).should.be.within(EXPECTED_META_CLASSES_MIN, EXPECTED_META_CLASSES_MAX);
 			});
 		});
 		it("should contain the snik subontology graphs", () => {
-			return sparql.select("select distinct(?g) {graph ?g {?class a owl:Class.}}").then((bindings) => {
+			return sparql.select("select distinct(?g) {graph ?g {?x a|rdfs:subClassOf meta:EntityType.}}").then((bindings) => {
 				const graphs = new Set();
 				for (const binding of bindings) {
 					graphs.add(binding.g.value);
 				}
-				const expectedGraphs = [
+				const expectedGraphs = new Set([
 					"http://www.snik.eu/ontology/meta",
 					"http://www.snik.eu/ontology/ob",
 					//'http://www.snik.eu/ontology/it',
@@ -39,9 +39,8 @@ describe("sparql", () => {
 					"http://www.snik.eu/ontology/bb",
 					"http://www.snik.eu/ontology/he",
 					"http://www.snik.eu/ontology/it4it",
-				];
-				const difference = new Set(expectedGraphs.filter((x) => !graphs.has(x))); //es6 set difference, see http://2ality.com/2015/01/es6-set-operations.html
-				assert.deepEqual([...difference], []);
+				]);
+				graphs.should.eql(expectedGraphs);
 			});
 		});
 		/*
