+++ js/config.dist.ts
@@ -5,7 +5,7 @@ Copy to js/config.ts after checkout and adapt to your preferences.
 import { LogLevelDesc } from "loglevel";
 
 const config = {
-	defaultSubOntologies: ["bb", "ob", "ciox", "he", "it4it"],
+	defaultSubOntologies: ["meta", "bb", "ob", "ciox", "he", "it4it"],
 	allSubOntologies: ["meta", "bb", "ob", "ciox", "he", "he-unconsolidated", "it4it"],
 	color: new Map([
 		["ciox", "rgb(80,255,250)"],
