### 2022-05-23 6dbc33d
Fix SPARQL loading.
