### 2022-05-25 cda207e
Resolve #350: Update dependencies, fix edge label autorotate. Set version to 22.05.
