### 2022-06-03 #169 @behollister
I'm open to doing the review. Please note that I am still finishing https://github.com/openjournals/joss-reviews/issues/4283.

Should be done with that one tomorrow, or the next day.
