### 2022-07-12 #180 @whedon
[REVIEW]: SNIK Graph—Visualizing Knowledge about Management of Hospital Information Systems
**Submitting author:** @KonradHoeffner (<a href="http://orcid.org/0000-0001-7358-3217">Konrad Höffner</a>)
**Repository:** <a href="https://github.com/snikproject/snik-graph" target ="_blank">https://github.com/snikproject/snik-graph</a>
**Version:** 3.0.0
