### 2022-07-19 #180 @labarba
Hi @bwatson, @behollister 👋 — Thank you for agreeing to review for JOSE! I am the editor-in-chief, and @juanklopper is the handling editor for this submission.

This issue thread is where the action happens: work your way through the review checklist, feel free to ask questions or post comments here, and also open issues in the submission repository as needed. Godspeed!
