### 2022-10-06 158b368
Simplify mass function. Should have no impact on #389.
