### 2022-10-06 cd452d4
Reduce simulation time and max iterations on tight layout. Part of #389.
