+++ js/layout.ts
@@ -283,8 +283,8 @@ export const eulerTight = {
 	name: "euler",
 	springLength: 40,
 	animate: false,
-	maxSimulationTime: 40000,
-	maxIterations: 500,
+	maxSimulationTime: 10000,
+	maxIterations: 100,
 	timeStep: 80,
 	refresh: 50,
 	randomize: false,
