+++ assets/main.5818b085.js
@@ -44,7 +44,7 @@ URL: ${a}
         owl:annotatedTarget ?d.
       }
     }`,s=await dr(o);return a.stop(s.length+" properties using SNIK query"),s}async function ad(t,e,r,n){const a=await nd(t,e,r),i=[];for(let o=0;o<a.length;o++)i.push({group:"edges",data:{source:a[o].c.value,target:a[o].d.value,id:String(o),p:a[o].p.value,pl:a[o].p.value.replace(/.*[#/]/,""),...a[o].g&&{g:a[o].g.value},...a[o].ax&&{ax:a[o].ax.value}}});return se.debug(a.length+" Edges loaded from SPARQL"),i}async function id(t,e){if(!e)return el(t);const[r,n]=await Promise.all([el(t),rd(t)]);return r.concat(n)}async function ac(t,e,r=!1,n=!1){se.debug(`Loading graph from endpoint ${Ae.sparql.endpoint} with graphs ${e}.`);const a=e.map(l=>`FROM <${l}>`).reduce((l,u)=>l+`
-`+u,""),i=a.replace(/FROM/g,"FROM NAMED"),[o,s]=await Promise.all([id(a,r),ad(a,i,r)]);t.elements().remove(),t.add(o),t.add(s),t.elements().addClass("unfiltered")}const xe={ID:"id",LABEL:"l",LABEL_GERMAN:"de",LABEL_ENGLISH:"en",LABEL_PERSIAN:"fa",SUBTOP:"st",SOURCE:"source",INSTANCE:"inst",SUBTOP_ENTITY_TYPE:"E",SUBTOP_ROLE:"R",SUBTOP_FUNCTION:"F"},od="22.05",Ur=od,sd=500;function It(t){const e=document.getElementById(t);if(!e)throw new Error(`Element with id ${t} does not exist.`);return e}const ld="https://github.com/snikproject/snik-graph",ui="https://github.com/snikproject/ontology";function ea(t,e,r,n){let a=encodeURIComponent(r);if(n){const i=n.map(s=>encodeURIComponent(s));let o=i.reduce((s,l)=>s+"%0A"+l);for(;o.length>sd;)i.shift(),o=i.reduce((s,l)=>s+"%0A"+l,"");a+="%0A%60%60%60%0A"+o+"%0A%60%60%60"}window.open(`${t}/issues/new?title=${encodeURIComponent(e)}&body=${a}`)}const ic=t=>e=>{switch(e.key){case" ":case"Enter":t.click()}};function oc(t,e,r){const n=document.createElement("div");return n.classList.add("dropdown-entry-checkboxtext"),n.innerText=e,r&&n.setAttribute("data-i18n",r),n.addEventListener("click",()=>{t.click()}),n}const Me={SOURCE:"source",TARGET:"target",ID:"id",PROPERTY:"p",PROPERTY_LABEL:"pl",GRAPH:"g",AXIOM:"ax"},ud={idStrings:{file:"File","load-sparql":"Load from SPARQL endpoint","load-session":"Load session","load-view":"Load Partial Graph","save-session":"Save Session","save-snik-graph":"Save complete SNIK Graph as Cytoscape File","save-view":"Save current Partial Graph","recalculate-layout-replace":"Recalculate Layout and Replace in Browser Cache","save-image-current-view":"Save Image of Current View","save-image-whole-graph":"Save Image of Whole Graph","save-image-current-view-high-res":"Save Image of Current View (high res)","save-image-whole-graph-high-res":"Save Image of Whole Graph (high res)",filter:"Filter",options:"Options",separateSubs:"separate subontologies",cumulativeSearch:"cumulative search",grid:"show grid",dayMode:"day mode",devMode:"develope
[…]
