+++ vite.config.js
@@ -4,6 +4,7 @@ import { fileURLToPath } from "url";
 const _dirname = typeof __dirname !== "undefined" ? __dirname : dirname(fileURLToPath(import.meta.url));
 
 export default {
+	base: "",
 	sourcemap: true,
 	build: {
 		rollupOptions: {
