+++ Dockerfile
@@ -7,8 +7,6 @@ RUN cp js/config.dist.ts js/config.ts
 ARG SPARQL_ENDPOINT=https://www.snik.eu/sparql
 RUN sed -i "s|https://www\.snik\.eu/sparql|${SPARQL_ENDPOINT}|" js/config.ts && \
 	npm run build && \
-	sed -i "s|/assets|./assets|" dist/index.html && \
-	sed -i "s|/assets|../assets|" dist/html/*.html && \
 	npm run doc
 
 FROM pierrezemb/gostatic
