### 2022-10-07 7134ba5
Show git commit info in about alert and GitHub issue template. Resolve #388.
