### 2022-10-10 433b44b
Generate sourcemap on build in GitHub Action.
