### 2022-10-10 b338fae
Enable source map option in package.json.
