### 2022-10-21 6ca5025
Layout (compound): Changed cose to grid mode to display as much nodes as possible. Added option for colored edges.
