+++ js/browser/util.ts
@@ -73,3 +73,18 @@ export function checkboxClickableDiv(box: HTMLInputElement, text: string, i18n: 
 	});
 	return div;
 }
+
+
+export function stringToColor(str: string) : string
+{
+    let hash = 0;
+    for (var i = 0; i < str.length; i++) {
+        hash = str.charCodeAt(i) + ((hash << 5) - hash);
+    }
+    let color = '#';
+    for (var i = 0; i < 3; i++) {
+        var value = (hash >> (i * 8)) & 0xFF;
+        color += ('00' + value.toString(16)).substr(-2);
+    }
+    return color;
+}
