### 2022-10-21 a2dc1c1
Added option for colored edges. Commit modified by Konrad Höffner.
