+++ assets/main.7492bed6.js
@@ -168,9 +168,9 @@ In order to be iterable, non-array objects must have a [Symbol.iterator]() metho
     ?c ?p ?o; a owl:Class.
     ?c meta:chapter/meta:subChapterOf* ?ch.
   } ORDER BY ASC(?ch)`,o=await dr(i);for(const s of o){const l=s.ch.value,u=r.insertRow(),c=u.insertCell();c.addEventListener("click",async()=>qu(t,await Hu(e,l)));const f=document.createElement("a");c.appendChild(f),f.innerText=s.ch.value;const d=u.insertCell();d.innerHTML=s.count.value;const v=u.insertCell(),h=document.createElement("input");v.appendChild(h),h.setAttribute("type","checkbox"),h.classList.add("chaptersearch-checkbox"),v.addEventListener("input",async()=>{h.checked?(await Hu(e,l),oi.add(l)):oi.delete(l),qu(t,new Set([...oi].map(p=>Th.get(p)).flat().sort()))})}}function Ys(t,e){const r=t.target.files[0],n=new FileReader;n.onload=()=>e(JSON.parse(n.result)),n.readAsText(r)}function Oi(t,e){const r=t.cy;r.elements().remove(),r.json(e),r.elements().addClass("unfiltered");const n=1*t.cy.elements(":visible").size()/t.cy.elements().size(),a=n<.8;se.debug("Load Graph from File: Visible fraction: "+n+" set star mode to "+a),t.cy.nodes(":child").size()>0&&(document.getElementById("combineMatchModeBox").checked=!0),t.starMode=a;const i=r.nodes(":visible");r.center(i),r.fit(i),r.elements().removeClass("highlighted"),r.elements().removeClass("source")}const RE=t=>e=>{Ys(e,r=>{Oi(t,r)})};async function NE(t){!confirm("This will override the current session. Continue?")||Ys(t,async e=>{if(e.state.version!==Gr&&!confirm(`Your file was saved in version ${e.state.version}, but SNIK Graph has version ${Gr}, so it may not work properly. Continue anyway?`))return;be.reset();const r=new be(!1),n=[r.initialized];Oi(r.state.graph,e.mainGraph.graph),be.activeView().setTitle(e.mainGraph.title);for(let a=0;a<e.tabs.length;a++){const i=new be(!1);n.push(i.initialized),Oi(i.state.graph,e.tabs[a].graph),be.activeView().setTitle(e.tabs[a].title)}await Promise.all(n),DE(e.state)})}function zE(t){Ys(t,e=>{if(e.version!==Gr&&!confirm(`Your file was saved in version ${e.version}, but SNIK Graph has version ${Gr}, so it may not work properly. Continue anyway?`))return;const r=new be(!1);Oi(r.state.graph,e.graph),be.activeView().setTitle(e.title)})}function Uu(t,e,r,n,a){const i=document.createElement("a");a.push(i),i.classList.add("dropdown-entry"),i.setAttribute("tabindex","-1"),t.appendChild(i);const o=document.createElement("input");o.type="file",o.style.display="none",i.appendChild(o);const s=document.createElement("span");s.innerText=r,s.setAttribute("data-i18n",e),i.appendChild(s),i.addEventListener("cl
[…]
