### 2022-10-24 a034648
Added hsv colors to colored edges. Added comments. (#392)
* Added hsv colors to colored edges. Added comments.
* Use prettier.

Co-authored-by: pc-guest <ralle_666@hotmail.de>
Co-authored-by: Konrad Höffner <konrad.hoeffner@uni-leipzig.de>
