### 2022-10-24 cb8688b
Fix @typescript-eslint/no-use-before-define linting error.
