+++ assets/main.85ca6ce2.js
@@ -44,7 +44,7 @@ URL: ${a}
         owl:annotatedTarget ?d.
       }
     }`,s=await dr(o);return a.stop(s.length+" properties using SNIK query"),s}async function ad(t,e,r,n){const a=await nd(t,e,r),i=[];for(let o=0;o<a.length;o++)i.push({group:"edges",data:{source:a[o].c.value,target:a[o].d.value,id:String(o),p:a[o].p.value,pl:a[o].p.value.replace(/.*[#/]/,""),...a[o].g&&{g:a[o].g.value},...a[o].ax&&{ax:a[o].ax.value}}});return se.debug(a.length+" Edges loaded from SPARQL"),i}async function id(t,e){if(!e)return el(t);const[r,n]=await Promise.all([el(t),rd(t)]);return r.concat(n)}async function ic(t,e,r=!1,n=!1){se.debug(`Loading graph from endpoint ${Ae.sparql.endpoint} with graphs ${e}.`);const a=e.map(l=>`FROM <${l}>`).reduce((l,u)=>l+`
-`+u,""),i=a.replace(/FROM/g,"FROM NAMED"),[o,s]=await Promise.all([id(a,r),ad(a,i,r)]);t.elements().remove(),t.add(o),t.add(s),t.elements().addClass("unfiltered")}const xe={ID:"id",LABEL:"l",LABEL_GERMAN:"de",LABEL_ENGLISH:"en",LABEL_PERSIAN:"fa",SUBTOP:"st",SOURCE:"source",INSTANCE:"inst",SUBTOP_ENTITY_TYPE:"E",SUBTOP_ROLE:"R",SUBTOP_FUNCTION:"F"},od="22.10",Gr=od,sd=500;function It(t){const e=document.getElementById(t);if(!e)throw new Error(`Element with id ${t} does not exist.`);return e}const ld="https://github.com/snikproject/snik-graph",ui="https://github.com/snikproject/ontology";function ta(t,e,r,n){let a=encodeURIComponent(r);if(n){const i=n.map(s=>encodeURIComponent(s.trim()));let o=i.reduce((s,l)=>s+"%0A"+l);for(;o.length>sd;)i.shift(),o=i.reduce((s,l)=>s+"%0A"+l,"");a+="%0A%0A%23%23%20Log%0A%60%60%60"+o+"%0A%60%60%60"}window.open(`${t}/issues/new?title=${encodeURIComponent(e)}&body=${a}`)}const oc=t=>e=>{switch(e.key){case" ":case"Enter":t.click()}};function sc(t,e,r){const n=document.createElement("div");return n.classList.add("dropdown-entry-checkboxtext"),n.innerText=e,r&&n.setAttribute("data-i18n",r),n.addEventListener("click",()=>{t.click()}),n}function ud(t){let e=0;for(let r=0;r<t.length;r++)e=t.charCodeAt(r)+((e<<5)-e);return e=e%180/360,e+=.5,cd(e,1,1)}function cd(t,e,r){const n=Math.floor(t*6),a=t*6-n,i=r*(1-e),o=r*(1-a*e),s=r*(1-(1-a)*e);let l,u,c;switch(n%6){case 0:l=r,u=s,c=i;break;case 1:l=o,u=r,c=i;break;case 2:l=i,u=r,c=s;break;case 3:l=i,u=o,c=r;break;case 4:l=s,u=i,c=r;break;case 5:l=r,u=i,c=o;break}return"#"+("00"+(l*255).toString(16)).slice(-2)+("00"+(u*255).toString(16)).slice(-2)+("00"+(c*255).toString(16)).slice(-2)}const Me={SOURCE:"source"
[…]
