### 2022-10-31 snikproject/graph#393 @behollister
Local Installation Instructions
Browser: https://www.whatismybrowser.com/w/ERBFHNF

From Node instructions:

_git clone https://github.com/snikproject/snik-graph.git
npm install
cp js/config.dist.ts js/config.ts
npm run preview_

**Appears to missing a 'cd snik-graph' after 'git clone' line to indicate that 'npm install' must be run out of the repository directory.**
