### 2022-11-01 726e84e
Fix npm install warning by using the newer --omit=dev instead of --production.
