### 2022-11-01 00d30bf
Add additional Dockerfile for when SNIK Graph is a submodule. Resolve #394.
