+++ README.md
@@ -38,6 +38,8 @@ Developing on Windows is not recommended but possible with the following adaptio
 
 Open <http://localhost:8043/> in a browser.
 
+When using SNIK Graph as a Git submodule, use `docker build -f Dockerfile.submodule .. -t snik-graph` instead, so that Git commit information can be included in the about dialogue, otherwise the build fails.
+
 ## Usage Requirements
 Optimized for PC with mouse, does not work well on mobile devices.
 Requires a browser with [ES6 module support](https://caniuse.com/es6-module).
