### 2022-11-01 snikproject/graph#393 @KonradHoeffner
Thank you for the correction! I added the `cd` statement to the documentation in commit b5a2ee34fd53fe514c08970420c4a9c13f696fc7.
