### 2022-11-01 snikproject/graph#395 @behollister
broken link under developer docs
Browser: https://whatismybrowser.com/w/9X2B92J

Broken link for "Developer Documentation"
Documentation
[Developer Documentation](https://snikproject.github.io/snik-graph/doc)
