### 2022-11-01 #180 @behollister
finished review. only would make suggestions about usability, such as scaled labels for nodes hard to see at various zoom levels, and overlapping popup text for menu items.

otherwise, all claims have been met by project for jose.
