### 2022-11-01 #180 @behollister
one more note. was not able to reach service locally on windows after having completed all installation steps in docs. worked fine on linux however.
