### 2022-11-02 #180 @KonradHoeffner
> Issue with developer docs. See [snikproject/graph#395 (comment)](https://github.com/snikproject/graph/issues/395#issue-1431896646)

The issue has been fixed, thanks for notifying us!
