### 2022-11-02 #180 @KonradHoeffner
> one more note. was not able to reach service locally on windows after having completed all installation steps in docs. worked fine on Linux however.

We usually develop under Linux but would like to enable development under Windows as well.
Can you share what happened exactly so we can extend our [Windows workaround](https://github.com/snikproject/graph/issues/379) section in the docs?
