### 2022-11-02 e61fe8f
Add reference to Thomas Pause's bachelor thesis.
