+++ paper/paper.bib
@@ -69,13 +69,13 @@
   publisher={Gesellschaft f{\"u}r Informatik eV}
 }
 
-@Article{sniktec,
-   Author="H{\"o}ffner, Konrad and Jahn, Franziska and K{\"u}cherer, Christian and Paech, Barbara and Brigel, Schneider and Sch{\"o}bel, Martin and St{\"a}ubert, Sebastian and Winter, Alfred",
-   Title="{{T}echnical {E}nvironment for {D}eveloping the {S}{N}{I}{K} {O}ntology of {I}nformation {M}anagement in {H}ospitals}",
-   Journal="Studies in Health Technology and Informatics",
-   Year="2017",
-   Volume="243",
-   Pages="122--126"
+@article{sniktec,
+  author="H{\"o}ffner, Konrad and Jahn, Franziska and K{\"u}cherer, Christian and Paech, Barbara and Brigel, Schneider and Sch{\"o}bel, Martin and St{\"a}ubert, Sebastian and Winter, Alfred",
+  title="{{T}echnical {E}nvironment for {D}eveloping the {S}{N}{I}{K} {O}ntology of {I}nformation {M}anagement in {H}ospitals}",
+  journal="Studies in Health Technology and Informatics",
+  year="2017",
+  volume="243",
+  pages="122--126"
 }
 
 @inproceedings{visualizationoflargeontologies,
@@ -135,4 +135,21 @@
   publisher = {GitHub},
   journal = {GitHub repository},
   url = {https://github.com/krisk/fuse}
+}
+
+@inproceedings{snikgraphposter,
+  title={The {SNIK} {G}raph: {V}isualization of a Medical Informatics Ontology},
+  author={Franziska Jahn and Konrad Höffner and Birgit Schneider and Anna Lörke and Thomas Pause and Elske Ammenwerth and Alfred Winter},
+  booktitle={MedInfo 2019, The 17th World Congress of Medical and Health Informatics, Lyon},
+  year={2019}
+}
+ 
+@mastersthesis{thomasba, 
+  title={{O}rganisation von {T}eilgraphen semantischer {N}etze zur ontologiegestützten {W}issensvermittlung}, 
+  author={Thomas Pause}, 
+  type={Bachelor's Thesis},
+  school={Institute for Medical Informatics, Statistics {and} Epidemiology (IMISE)},
+  address={Leipzig, Germany},
+  url={https://www.snik.eu/public/bachelor-tp.pdf},
+  year={2020} 
 }
