### 2022-11-03 2ba2626
Add comments to Direction.
