### 2022-11-08 ea458ac
Update dependencies including Vite. Should fix Windows parse5 bug and thus resolve #396.
