+++ assets/main.8c9c8804.js
@@ -183,9 +183,9 @@ In order to be iterable, non-array objects must have a [Symbol.iterator]() metho
     ?c ?p ?o; a owl:Class.
     ?c meta:chapter/meta:subChapterOf* ?ch.
   } ORDER BY ASC(?ch)`,o=await wr(a);for(const s of o){const l=s.ch.value,u=n.insertRow(),c=u.insertCell();c.addEventListener("click",async()=>dd(t,await fd(e,l)));const d=document.createElement("a");c.appendChild(d),d.innerText=s.ch.value;const p=u.insertCell();p.innerHTML=s.count.value;const m=u.insertCell(),h=document.createElement("input");m.appendChild(h),h.setAttribute("type","checkbox"),h.classList.add("chaptersearch-checkbox"),m.addEventListener("input",async()=>{h.checked?(await fd(e,l),Go.add(l)):Go.delete(l),dd(t,new Set([...Go].map(g=>Gv.get(g)).flat().sort()))})}}function Qu(t,e){const n=t.target.files[0],r=new FileReader;r.onload=()=>e(JSON.parse(r.result)),r.readAsText(n)}function ws(t,e){const n=t.cy;n.elements().remove(),n.json(e),n.elements().addClass("unfiltered");const r=1*t.cy.elements(":visible").size()/t.cy.elements().size(),i=r<.8;Me.debug("Load Graph from File: Visible fraction: "+r+" set star mode to "+i),t.cy.nodes(":child").size()>0&&(document.getElementById("combineMatchModeBox").checked=!0),t.starMode=i;const a=n.nodes(":visible");n.center(a),n.fit(a),n.elements().removeClass("highlighted"),n.elements().removeClass("source")}const BT=t=>e=>{Qu(e,n=>{ws(t,n)})};async function RT(t){!confirm("This will override the current session. Continue?")||Qu(t,async e=>{if(e.state.version!==si&&!confirm(`Your file was saved in version ${e.state.version}, but SNIK Graph has version ${si}, so it may not work properly. Continue anyway?`))return;Ve.reset();const n=new Ve(!1),r=[n.initialized];ws(n.state.graph,e.mainGraph.graph),Ve.activeView().setTitle(e.mainGraph.title);for(let i=0;i<e.tabs.length;i++){const a=new Ve(!1);r.push(a.initialized),ws(a.state.graph,e.tabs[i].graph),Ve.activeView().setTitle(e.tabs[i].title)}await Promise.all(r),MT(e.state)})}function zT(t){Qu(t,e=>{if(e.version!==si&&!confirm(`Your file was saved in version ${e.version}, but SNIK Graph has version ${si}, so it may not work properly. Continue anyway?`))return;const n=new Ve(!1);ws(n.state.graph,e.graph),Ve.activeView().setTitle(e.title)})}function vd(t,e,n,r,i){const a=document.createElement("a");i.push(a),a.classList.add("dropdown-entry"),a.setAttribute("tabindex","-1"),t.appendChild(a);const o=document.createElement("input");o.type="file",o.style.display="none",a.appendChild(o);const s=document.createElement("span");s.innerText=n,s.setAttribute("data-i18n
[…]
