### 2022-12-06 #180 @labarba
@KonradHoeffner — I request your patience in this, with a gentle reminder to leave the task of following up with reviewers to the editor. You are welcome to tag me or @juanklopper, but try not to ping the reviewer directly with reminders. This is just to respect editorial roles (remembering that everyone is a volunteer!). Of course, you can address the reviewers directly when responding to their review comments. Thanks!
