### 2023-03-01 #180 @KonradHoeffner
@behollister: The overlapping mouseover text was adressed in #379 and is now fixed in 974580af526186c33fead46447d3908e5b7ab5d9.
