### 2024-06-17 #180 @labarba
Hi @KonradHoeffner — thank you ever so much for your kind patience. I've read the paper and found two tiny typos: repetitions on page 3 ("of of") and on page 4 ("the the"). Do make these little fixes, then please make a tagged release for the publication and report the version number here. Finally, you should make an archive deposit on Zenodo and report the DOI here.
