### 2024-06-17 #180 @KonradHoeffner
@labarba: Thank you for the corrections! I fixed the typos and tagged a release for April 2024 (24.04) but I made some changes today that aren't exhaustively tested yet, so I will take a few days to think and then either report that as paper version or create a new release 24.06.
