### 2024-06-18 #180 @KonradHoeffner
@labarba: As requested I created a new release "24.06" with a tag of the same name and made a Zenodo archive deposit, the DOI is [10.5281/zenodo.12094738](https://doi.org/10.5281/zenodo.12094738).
