### 2024-06-19 #180 @KonradHoeffner
@labarba: I manually edited the author metadata of the zenodo release for version 24.06 to match the paper and will add a CITATION.cff so that it will be correct on further releases made by the Zenodo GitHub integration.
