### 2024-06-21 #180 @labarba
Gah! Now I see that the Zenodo archive shows a Creative Commons Attribution license, but the repository shows a GPL license. I think you need to change it in Zenodo to match!
