### 2024-06-24 #180 @KonradHoeffner
Sorry, I thought I changed this last week but I got some problems with Zenodo silently not accepting my metadata updates at that time, the license seems to stick now even after a reload.

![Screenshot from 2024-06-24 09-23-06](https://github.com/openjournals/jose-reviews/assets/839577/1ac29698-76b1-4687-8f44-55a40c41ec85)
