### 2024-06-24 #180 @KonradHoeffner
P.S.: I noticed that a student was missing that contributed to SNIK Graph in her Bachelor's thesis, so I added her to both the Zenodo meta data and the paper.
I also moved the leading senior researcher to the last authorship postion.
