### 2024-06-25 #180 @labarba
hmm... It's unusual to add to the author list this late in the process. Please give a few words about the change in authorship, and confirm that the added author was consulted, agreed to be listed as author, and stands by the material and publication.
