### 2024-06-26 #180 @KonradHoeffner
I haven't been able to reach the person because she was a student a long time ago and we don't have and also were not able to obtain current contact information.
So she wasn't actually involved in the paper however I remembered that her bachelor's thesis included significant theoretical contributions to SNIK Graph in the form of requirements which we collected as issues in the milestones <https://github.com/snikproject/graph/milestone/12> and <https://github.com/snikproject/graph/milestone/13> (surname changed since then) and the others implemented them partially.
As this paper does not have a political or otherwise controversial topic I considered that the ethical positives of honoring her contributions have a higher value than the problem of not being able to contact her for agreement.

However if you assess this differently it is no problem for me to remove the author again.
