### 2024-06-27 #180 @labarba
Under the circumstances you describe, I would suggest that you add this student in the acknowledgements, instead of a last-minute author addition without their explicit consent.
