### 2024-07-08 #180 @KonradHoeffner
@labarba: I have finally managed to contact the former student and she agrees that her bachelor thesis was a significant contribution that qualifiers for authorship and she consents to be added to the paper as coauthor, however if this is not possible so late in the review she also understands.
Is it OK if I add her now that the circumstances have changed? I would update the metadata of course so that it fits with the paper again.
