### 2024-07-09 #180 @labarba
I'm leaving this issue open for the time being because the PDF of the article is not showing yet. I am checking with @openjournals/dev whether something needs a fix there... Stay tuned.
