### 2024-07-10 #180 @KonradHoeffner
Thank you @labarba @juanklopper @behollister!

The DOI https://doi.org/10.21105/jose.00180 indeed sends me to the JOSE page of the paper https://jose.theoj.org/papers/10.21105/jose.00180 and the PDF is shown on this page as well for me.
